API/rest_menu_bot/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
idna==2.6
pytz==2018.4
requests==2.18.4
urllib3==1.22
